server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "alembic>=1.17.2",
    "cryptography>=46.0.3",
    "fastapi>=0.127.1",
    "fastapi-limiter>=0.1.6",
    "fastembed>=0.7.4",
    "passlib>=1.7.4",
    "psycopg>=3.3.2",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pydantic[email]>=2.12.5",
    "pyjwt>=2.10.1",
    "qdrant-client>=1.16.2",
    "redis>=7.1.0",
    "sqlalchemy>=2.0.45",
    "sqlmodel>=0.0.31",
    "uvicorn>=0.40.0",
    "youtube-transcript-api>=1.2.3",
]
